IP-470/Pr comments
