+++ lib/roo_on_rails/railties/sidekiq.rb
@@ -4,12 +4,19 @@ module RooOnRails
       initializer 'roo_on_rails.sidekiq' do |app|
         require 'hirefire-resource'
         $stderr.puts 'initializer roo_on_rails.sidekiq'
-        break unless ENV.fetch('SIDEKIQ_ENBALED', 'true').to_s =~ /\A(YES|TRUE|ON|1)\Z/i
+        break unless ENV.fetch('SIDEKIQ_ENABLED', 'true').to_s =~ /\A(YES|TRUE|ON|1)\Z/i
         config_hirefire(app)
       end
 
       def config_hirefire(app)
-        return unless ENV['HIREFIRE_TOKEN']
+        unless ENV['HIREFIRE_TOKEN']
+          warn 'No HIREFIRE_TOKEN token set, auto scaling not enabled'
+          return
+        end
+        add_middleware(app)
+      end
+
+      def add_middleware(app)
         $stderr.puts 'HIREFIRE_TOKEN set'
         app.middleware.use HireFire::Middleware
         HireFire::Resource.configure do |config|
