+++ roo_on_rails.gemspec
@@ -39,5 +39,4 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'simplecov'
   spec.add_development_dependency 'codecov'
   spec.add_development_dependency 'sidekiq'
-
 end
